data/requirements.txt
absl-py==1.4.0
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
Automat==22.10.0
Babel==2.15.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
blinker==1.8.2
Boruta==0.3
cachetools==5.3.0
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
constantly==15.1.0
construct==2.10.68
contourpy==1.2.1
cramjam==2.8.3
cycler==0.11.0
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dnspython==2.6.1
emlearn==0.21.0
exceptiongroup==1.2.1
executing==2.0.1
fastjsonschema==2.19.1
fastparquet==2024.2.0
filelock==3.13.1
Flask==3.0.3
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.29.1
fqdn==1.5.1
fsspec==2024.5.0
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
grpcio==1.54.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httpx==0.27.0
hyperlink==21.0.0
idna==3.4
imbalanced-learn==0.11.0
imblearn==0.0
incremental==22.10.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.8
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
keras==3.4.1
kiwisolver==1.3.2
libclang==16.0.0
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
mosquito==0.3.2
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.7.0.72
opt-einsum==3.3.0
optree==0.12.1
overrides==7.7.0
packaging==24.0
paho-mqtt==2.1.0
pandas==2.2.2
pandocfilters==1.5.1
parso==0.8.4
Pillow==9.0.1
platformdirs==4.1.0
ply==3.11
polars==0.19.19
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.22.4
psutil==5.9.8
pure-eval==0.2.2
pyarrow==16.1.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybind11==2.11.1
pycparser==2.22
pycryptodomex==3.17
pydot==1.4.2
pygeoip==0.3.2
Pygments==2.18.0
pyparsing==3.0.7
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
python-engineio==4.9.1
python-json-logger==2.0.7
python-socketio==5.11.2
pytz==2021.3
pywin32==306
pywinpty==2.0.13
PyYAML==6.0
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.13.2
Send2Trash==1.8.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
style==1.1.0
tensor==0.3.6
tensorboard==2.17.0
tensorboard-data-server==0.7.0
tensorflow==2.17.0
tensorflow-intel==2.17.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.1
threadpoolctl==3.1.0
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.3
Twisted==22.10.0
twisted-iocpsupport==1.0.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
update==0.0.1
uri-template==1.3.0
urllib3==2.0.2
virtualenv==20.25.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
widgetsnbextension==4.0.10
wrapt==1.14.1
wsproto==1.2.0
zope.interface==6.0
